Fix the warnings building the suites turned up
Most were comparisons between a signed and an unsigned expression, and in
nearly every one the test held the wrong type rather than the comparison being
awkward:

- the bit counting tests took expected as the unsigned type of the input,
  where bit_width and the count_* functions all return int;
- the slice tests cast the expectation to size_t, though slice::start_type is
  signed, so the cast was the wrong way round;
- the joint layout tests multiplied a size_t index by a ptrdiff_t stride and
  compared the unsigned result against a signed offset, and took the index
  GENERATE hands them as int where the cursor reports size_t;
- the byte order test compared the reversal of an int64_t against a literal
  too wide to be one.

The rest:

- a bool sum was written as a multiplication of two bools, which gives the
  right answer and reads as a mistaken &&. A sum of booleans saturates, so say
  that instead.
- 0xAA does not fit in a signed char, so allocate the scratch buffer as
  unsigned char.
- an operator== on deduce_outcome was never called; the tests compare the
  members.
- inner_index was read only from a pack expansion, which is empty when there
  are no operands, so the innermost axis is now indexed directly.

trompeloeil's own headers warn too, and are not ours to fix, so they are held
as system includes. A subdirectory build leaves them ordinary.

The build is now clean, warnings and all.

+++ tests/unitary/src/core/layout/test_index.cpp
@@ -15,7 +15,7 @@ TEST_CASE("sanitize_index with a in-bounds positive index should return the same
 	const std::ptrdiff_t index = GENERATE(0, 4, 8, 15);
 
 	const auto sanitized = sanitize_index(index, extent);
-	REQUIRE( sanitized == index );
+	REQUIRE( sanitized == static_cast<std::size_t>(index) );
 }
 
 TEST_CASE("sanitize_index with a in-bounds negative index should return the poisitivized value", "[index]")
